// desktop/src/open-api/model/searchResult.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReceiptStatus } from './receiptStatus';


export interface SearchResult { 
    id: number;
    name: string;
    type: string;
    groupId: number;
    date: string;
    amount?: string;
    receiptStatus?: ReceiptStatus;
    paidByUserId?: number;
    createdAt: string;
}
export namespace SearchResult {
}